SFO Airport Air Traffic Passenger Statistics

Monthly summary of number of passengers in San Francisco International Airport (SFO)

sfo_passengers

Format

A data frame with 12 variables.

activity_period

Activity year and month in YYYYMM format

operating_airline

Airline name for the operator of aircraft

operating_airline_iata_code

The International Air Transport Association (IATA) two-letter designation for the Operating Airline

published_airline

Airline name that issues the ticket and books revenue for passenger activity

published_airline_iata_code

The International Air Transport Association (IATA) two-letter designation for the Published Airline

geo_summary

Designates whether the passenger activity in relation to SFO arrived from or departed to a location within the United States (“domestic”), or outside the United States (“international”) without stops

geo_region

Provides a more detailed breakdown of the GEO Summary field to designate the region in the world where activity in relation to SFO arrived from or departed to without stops

activity_type_code

A description of the physical action a passenger took in relation to a flight, which includes boarding a flight (“enplanements”), getting off a flight (“deplanements”) and transiting to another location (“intransit”)

price_category_code

A categorization of whether a Published Airline is a low-cost carrier or not a low-cost carrier

terminal

The airport terminal designations at SFO where passenger activity took place

boarding_area

The airport boarding area designations at SFO where passenger activity took place

passenger_count

The number of monthly passengers associated with the above attribute fields

Source

San Francisco data portal (DataSF) website.

Details

The dataset contains the monthly summary of number of passengers in San Francisco International Airport (SFO)

Examples

data(sfo_passengers)

require(dplyr)
#> Loading required package: dplyr
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Get summary of total number of passengers by activity type
# in most recent month
sfo_passengers %>%
  filter(activity_period == max(activity_period)) %>%
  group_by(activity_type_code) %>%
  summarise(total = sum(passenger_count), .groups = "drop")
#> # A tibble: 3 × 2
#>   activity_type_code   total
#>   <chr>                <int>
#> 1 Deplaned           1830374
#> 2 Enplaned           1907342
#> 3 Thru / Transit        8442

# Get summary of total number of passengers by
# activity type and geo region in most recent month
sfo_passengers %>%
filter(activity_period == max(activity_period)) %>%
  group_by(activity_type_code, geo_region) %>%
  summarise(total = sum(passenger_count), .groups = "drop")
#> # A tibble: 19 × 3
#>    activity_type_code geo_region            total
#>    <chr>              <chr>                 <int>
#>  1 Deplaned           Asia                 147461
#>  2 Deplaned           Australia / Oceania   45950
#>  3 Deplaned           Canada                53237
#>  4 Deplaned           Central America       12749
#>  5 Deplaned           Europe               124301
#>  6 Deplaned           Mexico                64121
#>  7 Deplaned           Middle East           28921
#>  8 Deplaned           US                  1353634
#>  9 Enplaned           Asia                 166684
#> 10 Enplaned           Australia / Oceania   44286
#> 11 Enplaned           Canada                57280
#> 12 Enplaned           Central America       15475
#> 13 Enplaned           Europe               135635
#> 14 Enplaned           Mexico                73971
#> 15 Enplaned           Middle East           31446
#> 16 Enplaned           US                  1382565
#> 17 Thru / Transit     Australia / Oceania    2817
#> 18 Thru / Transit     Europe                 2496
#> 19 Thru / Transit     US                     3129