SFO Airport Air Landings Statistics

Monthly statistics on San Francisco International Airport (SFO) landings

sfo_stats

Format

A data frame with 14 variables.

activity_period

Activity year and month in YYYYMM format

operating_airline

Airline name for the operator of aircraft

operating_airline_iata_code

The International Air Transport Association (IATA) two-letter designation for the Operating Airline

published_airline

Airline name that issues the ticket and books revenue for passenger activity

published_airline_iata_code

The International Air Transport Association (IATA) two-letter designation for the Published Airline

geo_summary

Designates whether the passenger activity in relation to SFO arrived from or departed to a location within the United States (“domestic”), or outside the United States (“international”) without stops

geo_region

Provides a more detailed breakdown of the GEO Summary field to designate the region in the world where activity in relation to SFO arrived from or departed to without stops

landing_aircraft_type

A designation for three types of aircraft that landed at SFO, which includes passenger aircraft, cargo only aircraft (“freighters”) or combination aircraft (“combi”)

aircraft_body_type

A designation that is independent from Landing Aircraft Type, which determines whether commercial aircraft landed at SFO is a wide body jet, narrow body jet, regional jet or a propeller operated aircraft

aircraft_manufacturer

Manufacturer name for the aircraft that landed at SFO

aircraft_model

Model designation of aircraft by the manufacturer

aircraft_version

Variations of the Aircraft Model, also known as the “dash number”, designated by the manufacturer to segregate unique versions of the same model

landing_count

The number of aircraft landings associated with General and Landings Statistics attribute fields

total_landed_weight

The aircraft landed weight (in pounds) associated with General and Landings Statistics attribute fields

Source

San Francisco data portal (DataSF) website.

Details

The dataset contains the monthly statistics on the air traffic landings in San Francisco International Airport (SFO)

Examples

data(sfo_stats)

require(dplyr)

# Get summary of total landing and weight by geo region
# in most recent month
sfo_stats %>%
  filter(activity_period == max(activity_period)) %>%
  group_by(geo_region) %>%
  summarise(total_landing = sum(landing_count),
            total_weight = sum(total_landed_weight),
  .groups = "drop")

#> # A tibble: 8 x 3
#>   geo_region          total_landing total_weight
#>   <chr>                       <int>        <int>
#> 1 Asia                          404    207684731
#> 2 Australia / Oceania            74     32992000
#> 3 Canada                         92      9022707
#> 4 Central America                 8      1102312
#> 5 Europe                        175     82551871
#> 6 Mexico                        137     19783165
#> 7 Middle East                    12      5120000
#> 8 US                           6238    857513800